change namefile of ToF implementation and add a IR sensor version
